Initial commit of PCAP parser plugin
Using PacketFu this time around to parse PCAP data. My hope is that by
using PacketFu, we can create custom protocol representations in the
PacketFu library to determine if a packet is for a particular protocol
of interest (i.e. create PacketFu protocol parsers for each control
system protocol we're interested in).

The other option for detecting protocols is to develop a filter
capability that packet parsers run packets through each time they parse
traffic.

Right now, this plugin is looking at the MAC addresses and creating an
ethernet interface and associated IP interface if the MAC address
isn't found. Need to improve on this by first checking to see if the IP
interface already exists, and if so updating it to include the MAC
address defined in the packet. There may be cases in which the IP
interface already has an Ethernet interface. If it does and the MAC
address doesn't match the one on the wire, we'll need to updated based
on certainty factors.