Inital commit of support for plugins as Rubygems
Leveraging the built-in post_install/uninstall hooks in the Rubygems
framework to copy ANTFARM plugin files from the Rubygem package to the
user's home directory on install and delete the plugin files on
uninstall.

This is by way of the `lib/rubygems_plugin.rb` file, which the Rubygems
framework loads each time the `gem` command is invoked. The
install/uninstall hooks are setup to look for any gems named
`antfarm-xyz` and assume the name of the plugin is therefore `xyz`. A
directory is then created in the user's plugins directory with the name
`xyz` and any Ruby and man files are copied over on install and deleted
on uninstall.