Overturning the last commit due to it breaking the main binary
Plugins need to be loaded during the init stage so they can be added to
Trollop subcommands, but the environment can't be processed until after
Trollop has run and grabbed global options. So, rather than being able
to specify a custom plugin directory, I just fixed the Dir globs in the
plugin loader to work with symlinks.